requirements.txt
discordpy
mcrcon